+++ components/frontend/src/metric/MetricDetails.js
@@ -17,7 +17,7 @@ import { MetricConfigurationParameters } from './MetricConfigurationParameters';
 import { MetricDebtParameters } from './MetricDebtParameters';
 import { MetricTypeHeader } from './MetricTypeHeader';
 import { TrendGraph } from './TrendGraph';
-import { datePropType, reportPropType, reportsPropType, stringsPropType, stringsURLSearchQueryPropType} from '../sharedPropTypes';
+import { datePropType, reportPropType, reportsPropType, settingsPropType, stringsPropType, stringsURLSearchQueryPropType } from '../sharedPropTypes';
 
 function Buttons({ isFirstMetric, isLastMetric, metric_uuid, reload, stopSorting }) {
     return (
@@ -53,7 +53,6 @@ fetchMeasurements.propTypes = {
     setMeasurements: PropTypes.func
 }
 
-
 export function MetricDetails({
     changed_fields,
     isFirstMetric,
@@ -63,6 +62,7 @@ export function MetricDetails({
     report_date,
     reports,
     report,
+    settings,
     stopSorting,
     subject_uuid,
     visibleDetailsTabs,
@@ -140,7 +140,7 @@ export function MetricDetails({
             const source_name = get_source_name(report_source, dataModel);
             panes.push({
                 menuItem: <Menu.Item key={source.source_uuid}><FocusableTab>{source_name}</FocusableTab></Menu.Item>,
-                render: () => <Tab.Pane><SourceEntities report={report} metric={metric} metric_uuid={metric_uuid} source={source} reload={measurementsReload} /></Tab.Pane>
+                render: () => <Tab.Pane><SourceEntities report={report} metric={metric} metric_uuid={metric_uuid} source={source} reload={measurementsReload} settings={settings} /></Tab.Pane>
             });
         });
     }
@@ -172,6 +172,7 @@ MetricDetails.propTypes = {
     report_date: datePropType,
     reports: reportsPropType,
     report: reportPropType,
+    settings: settingsPropType,
     stopSorting: PropTypes.func,
     subject_uuid: PropTypes.string,
     visibleDetailsTabs: stringsURLSearchQueryPropType