+++ components/frontend/src/source/SourceEntityDetails.test.js
@@ -15,9 +15,11 @@ function renderSourceEntityDetails() {
     render(
         <Permissions.Provider value={[EDIT_ENTITY_PERMISSION]}>
             <SourceEntityDetails
+                metric={{}}
                 metric_uuid="metric_uuid"
                 source_uuid="source_uuid"
                 entity={{ key: "key" }}
+                report={{ report_uuid: "report_uuid" }}
                 status="unconfirmed"
                 name="violation"
                 reload={reload}
@@ -38,7 +40,7 @@ it('changes the entity status', () => {
 it('changes the entity status end date', async () => {
     source.set_source_entity_attribute = jest.fn()
     renderSourceEntityDetails()
-    await userEvent.type(screen.getByPlaceholderText(/YYYY-MM-DD/), '2222-01-01{Tab}', {initialSelectionStart: 0, initialSelectionEnd: 10} );
+    await userEvent.type(screen.getByPlaceholderText(/YYYY-MM-DD/), '2222-01-01{Tab}', { initialSelectionStart: 0, initialSelectionEnd: 10 });
     expect(source.set_source_entity_attribute).toHaveBeenCalledWith(
         "metric_uuid", "source_uuid", "key", "status_end_date", "2222-01-01", reload
     );